remove src folder, restrcuture tests to include conftest and remove uneeded test class
